# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API PU-SDA-JATIM",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/payment/create": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "operationId": "paymentadd",
                "parameters": [
                    {
                        "name": "childrens_id",
                        "in": "query",
                        "description": "childrens_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "year ",
                        "in": "query",
                        "description": "year ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "proof_of_payment",
                        "in": "query",
                        "description": "proof_of_payment",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "payment_amount",
                        "in": "query",
                        "description": "payment_amount",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/payment/update/{id}": {
            "put": {
                "tags": [
                    "Payments"
                ],
                "operationId": "paymentupdate",
                "parameters": [
                    {
                        "name": "childrens_id",
                        "in": "query",
                        "description": "childrens_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "year ",
                        "in": "query",
                        "description": "year ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "proof_of_payment",
                        "in": "query",
                        "description": "proof_of_payment",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "payment_amount",
                        "in": "query",
                        "description": "payment_amount",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    }
}